Remove obsolete APIResponsiveness measurement.
